Fetch default branch for the branch buttons
If not available on page
